feat: move apply to client and assign labels when applying
