+++ pkg/template/processor_test.go
@@ -2,8 +2,6 @@ package template_test
 
 import (
 	"bytes"
-	"context"
-	"errors"
 	"fmt"
 	"testing"
 	texttemplate "text/template"
@@ -16,15 +14,10 @@ import (
 	authv1 "github.com/openshift/api/authorization/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func TestProcess(t *testing.T) {
@@ -37,8 +30,7 @@ func TestProcess(t *testing.T) {
 	codecFactory := serializer.NewCodecFactory(s)
 	decoder := codecFactory.UniversalDeserializer()
 
-	cl := NewFakeClient(t)
-	p := template.NewProcessor(cl, s)
+	p := template.NewProcessor(s)
 
 	t.Run("should process template successfully", func(t *testing.T) {
 		// given
@@ -224,256 +216,6 @@ func TestProcess(t *testing.T) {
 	})
 }
 
-func TestProcessAndApply(t *testing.T) {
-
-	commit := getNameWithTimestamp("sha")
-	user := getNameWithTimestamp("user")
-
-	s := addToScheme(t)
-	codecFactory := serializer.NewCodecFactory(s)
-	decoder := codecFactory.UniversalDeserializer()
-
-	values := map[string]string{
-		"USERNAME": user,
-		"COMMIT":   commit,
-	}
-
-	t.Run("should create namespace alone", func(t *testing.T) {
-		// given
-		cl := NewFakeClient(t)
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(Namespace), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-
-		// when
-		createdOrUpdated, err := p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		assertNamespaceExists(t, cl, user)
-	})
-
-	t.Run("should create role binding alone", func(t *testing.T) {
-		// given
-		cl := NewFakeClient(t)
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(RoleBinding), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-
-		// when
-		createdOrUpdated, err := p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		assertRoleBindingExists(t, cl, user)
-	})
-
-	t.Run("should create namespace and role binding", func(t *testing.T) {
-		// given
-		cl := NewFakeClient(t)
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(Namespace, RoleBinding), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-
-		// when
-		createdOrUpdated, err := p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		assertNamespaceExists(t, cl, user)
-		assertRoleBindingExists(t, cl, user)
-	})
-
-	t.Run("should update existing role binding", func(t *testing.T) {
-		// given
-		cl := NewFakeClient(t)
-		cl.MockCreate = func(ctx context.Context, obj runtime.Object, opts ...client.CreateOption) error {
-			meta, err := meta.Accessor(obj)
-			require.NoError(t, err)
-			meta.SetResourceVersion("1")
-			return cl.Client.Create(ctx, obj, opts...)
-		}
-		cl.MockUpdate = func(ctx context.Context, obj runtime.Object, opts ...client.UpdateOption) error {
-			meta, err := meta.Accessor(obj)
-			require.NoError(t, err)
-			t.Logf("updating resource of kind %s with version %s\n", obj.GetObjectKind().GroupVersionKind().Kind, meta.GetResourceVersion())
-			if obj.GetObjectKind().GroupVersionKind().Kind == "RoleBinding" && meta.GetResourceVersion() != "1" {
-				return fmt.Errorf("invalid resource version: %q", meta.GetResourceVersion())
-			}
-			return cl.Client.Update(ctx, obj, opts...)
-		}
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(RoleBinding), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-		createdOrUpdated, err := p.Apply(objs)
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		assertRoleBindingExists(t, cl, user)
-
-		// when rolebinding changes
-		tmpl, err = DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(Namespace, RoleBindingWithExtraUser), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err = p.Process(tmpl, values)
-		require.NoError(t, err)
-		createdOrUpdated, err = p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		binding := assertRoleBindingExists(t, cl, user)
-		require.Len(t, binding.Subjects, 2)
-		assert.Equal(t, "User", binding.Subjects[0].Kind)
-		assert.Equal(t, user, binding.Subjects[0].Name)
-		assert.Equal(t, "User", binding.Subjects[1].Kind)
-		assert.Equal(t, "extraUser", binding.Subjects[1].Name)
-	})
-
-	t.Run("should not create or update existing namespace and role binding", func(t *testing.T) {
-		// given
-		cl := NewFakeClient(t)
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(Namespace, RoleBinding), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-		created, err := p.Apply(objs)
-		require.NoError(t, err)
-		assert.True(t, created)
-		assertNamespaceExists(t, cl, user)
-		assertRoleBindingExists(t, cl, user)
-
-		// when apply the same template again
-		updated, err := p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.False(t, updated)
-	})
-
-	t.Run("failures", func(t *testing.T) {
-
-		t.Run("should fail to create template object", func(t *testing.T) {
-			// given
-			cl := NewFakeClient(t)
-			p := template.NewProcessor(cl, s)
-			cl.MockCreate = func(ctx context.Context, obj runtime.Object, opts ...client.CreateOption) error {
-				return errors.New("failed to create resource")
-			}
-			tmpl, err := DecodeTemplate(decoder,
-				CreateTemplate(WithObjects(RoleBinding), WithParams(UsernameParam, CommitParam)))
-			require.NoError(t, err)
-
-			// when
-			objs, err := p.Process(tmpl, values)
-			require.NoError(t, err)
-			createdOrUpdated, err := p.Apply(objs)
-
-			// then
-			require.Error(t, err)
-			assert.False(t, createdOrUpdated)
-		})
-
-		t.Run("should fail to update template object", func(t *testing.T) {
-			// given
-			cl := NewFakeClient(t)
-			p := template.NewProcessor(cl, s)
-			cl.MockUpdate = func(ctx context.Context, obj runtime.Object, opts ...client.UpdateOption) error {
-				return errors.New("failed to update resource")
-			}
-			tmpl, err := DecodeTemplate(decoder,
-				CreateTemplate(WithObjects(RoleBinding), WithParams(UsernameParam, CommitParam)))
-			require.NoError(t, err)
-			objs, err := p.Process(tmpl, values)
-			require.NoError(t, err)
-			createdOrUpdated, err := p.Apply(objs)
-			require.NoError(t, err)
-			assert.True(t, createdOrUpdated)
-
-			// when
-			tmpl, err = DecodeTemplate(decoder,
-				CreateTemplate(WithObjects(RoleBindingWithExtraUser), WithParams(UsernameParam, CommitParam)))
-			require.NoError(t, err)
-			objs, err = p.Process(tmpl, values)
-			require.NoError(t, err)
-			createdOrUpdated, err = p.Apply(objs)
-
-			// then
-			assert.Error(t, err)
-			assert.False(t, createdOrUpdated)
-		})
-	})
-
-	t.Run("should create with extra labels and ownerref", func(t *testing.T) {
-
-		// given
-		values := map[string]string{
-			"USERNAME": user,
-			"COMMIT":   commit,
-		}
-		cl := NewFakeClient(t)
-		p := template.NewProcessor(cl, s)
-		tmpl, err := DecodeTemplate(decoder,
-			CreateTemplate(WithObjects(Namespace, RoleBinding), WithParams(UsernameParam, CommitParam)))
-		require.NoError(t, err)
-		objs, err := p.Process(tmpl, values)
-		require.NoError(t, err)
-
-		// when adding labels and an owner reference
-		obj := objs[0]
-		meta, err := meta.Accessor(obj.Object)
-		require.NoError(t, err)
-		meta.SetOwnerReferences([]metav1.OwnerReference{
-			{
-				APIVersion: "crt/v1",
-				Kind:       "NSTemplateSet",
-				Name:       "foo",
-			},
-		})
-		meta.SetLabels(map[string]string{
-			"provider": "codeready-toolchain",
-			"version":  commit,
-			"extra":    "foo",
-		})
-		createdOrUpdated, err := p.Apply(objs)
-
-		// then
-		require.NoError(t, err)
-		assert.True(t, createdOrUpdated)
-		ns := assertNamespaceExists(t, cl, user)
-		// verify labels
-		assert.Equal(t, map[string]string{
-			"provider": "codeready-toolchain",
-			"version":  commit,
-			"extra":    "foo",
-		}, ns.Labels)
-		// verify owner refs
-		assert.Equal(t, []metav1.OwnerReference{
-			{
-				APIVersion: "crt/v1",
-				Kind:       "NSTemplateSet",
-				Name:       "foo",
-			},
-		}, ns.OwnerReferences)
-	})
-}
-
 func addToScheme(t *testing.T) *runtime.Scheme {
 	s := scheme.Scheme
 	err := authv1.Install(s)
@@ -521,25 +263,4 @@ func newObject(template, username, commit string) (runtime.Unstructured, error) 
 	result := &unstructured.Unstructured{}
 	err = result.UnmarshalJSON(buf.Bytes())
 	return result, err
-}
-
-func assertNamespaceExists(t *testing.T, c client.Client, nsName string) corev1.Namespace {
-	// check that the namespace was created
-	var ns corev1.Namespace
-	err := c.Get(context.TODO(), types.NamespacedName{Name: nsName, Namespace: ""}, &ns) // assert namespace is cluster-scoped
-	require.NoError(t, err)
-	return ns
-}
-
-func assertRoleBindingExists(t *testing.T, c client.Client, ns string) authv1.RoleBinding {
-	// check that the rolebinding is created in the namespace
-	// (the fake client just records the request but does not perform any consistency check)
-	var rb authv1.RoleBinding
-	err := c.Get(context.TODO(), types.NamespacedName{
-		Namespace: ns,
-		Name:      fmt.Sprintf("%s-edit", ns),
-	}, &rb)
-
-	require.NoError(t, err)
-	return rb
 }
